Move cli func to a dedicated module
